Initial draft of automatic CircleCI installation
Switch to from yawn-yaml to the underlying js-yaml dependency. The
higher level abstractions were making mutating the config awkward, it
had quite a few bugs, and we don't care so much about preserving the
comments and indentation when it's an automatically generated config
that is written out.

For now just the job names are written out, though we will also want to
include the filters and requires properties too.