Don't trigger new build after orb dev workflow
The trigger-integration-tests-workflow job in the orb-test-pack workflow
was telling CircleCI to trigger another round of workflows (via curl)
with the run-integration-tests variable enabled so as not to run the
same workflow, and to run orb-integration-test_deploy instead.
Unfortunately, the build-test workflow was running regardless, and the
orb-test-pack workflow was not being triggered when a new git tag was
created, so build-test was being run twice after every PR commit, and
orb-integration-test_deploy was never being run.

Not sure why CircleCI recommends using this approach rather than just
using the requires and filter properties that are much simpler.