Avoid duplicate jobs in generated CircleCI config
This was happening when a new version of tool-kit was installed, causing
the circleci plugin check to fail as the version in the config wouldn't
match. Previously, the install step would unconditionally add the
current job to the workflow's array. Now it checks for the job's
presence in the array first. Can't think of a nicer way around this
problem for now...