Set global option registry before checking install
Installation validation was previously failing if options used in
installation (such as the globs lint-staged hooks are installed for)
were overridden, as those changes wouldn't be read by the time the
check was run, so you'd get false negatives based on the default values.

Should be okay to split this logic out of the validation step as it was
optional already, so no invariants are violated.