Validate that plugins are valid when loading them
We do this by defining an abstract Hook class that all hooks must be
derived from. This means we can now check all the types on packages we
load have the appropriate instanceof, enforcing a structure for plugins
that should avoid us running into hard-to-debug issues down the road.

Any ideas for better names for the 'hook' package are welcome: it
defines both the Hook and Plugin types as they are mutually dependent!