Treat duplicate nightly jobs separately
Sometimes a job might exist in the main tool-kit workflow but not in the
nightly workflow so we need to check for duplicates in each separately.

Also avoid the need for deep cloning the job we want to push by
overriding the requires field during the cloning instead of reassigning
the property afterwards.