Organise packages into appropriate directories
Now all packages which define Tool Kit plugins, i.e., packages which
export hooks or tasks and are consumed by other plugins or apps in their
`plugins` fields are located in the `plugins` directory.
All libraries that other packages pull in but are not Tool Kit plugins
are located in the `lib` directory.
All other packages remain in the `packages` directory.

These changes will help us quickly differentiate between plugins and
helper libraries, both as developers and to allow us to run integration
tests on only the plugins.