Run future jobs even if approval is filtered out
We want all jobs to block on the `waiting-for-approval` job, however
this job only runs conditionally and if it is filtered out then the
`tool-kit/setup` job is consequently filtered out (and all other jobs
after it) as it's the only job it requires. Let's hack around this by
creating a very simple job which just does the `checkout` step that was
previously the first step of `tool-kit/setup` (which is quite simple and
requires few dependencies.) This job runs unconditionally and will allow
`tool-kit/setup` to run even if `waiting-for-approval` is filtered out.