Add special prompt to enter Heroku scaling config
It is difficult to create a generic way of prompting the user to give
settings for more complex options structures like the ones added in the
Heroku plugin, so instead we should allow plugins to create their own
prompt flows for specific options if necessary. We pass the prompt
function to them and they can construct a flow specific to their use
case, returning the resultant options object.

In this instance, the Heroku plugin now prompts the user to enter the
scaling configuration for each app in their Heroku pipeline.