Lazily get lint-staged-npm keys
lint-staged-npm allows you to set the globs used by lint-staged in your
.toolkitrc.yml config. Previously, it would read the options when the
class was constructed. However, now that we construct the classes when
first loading the plugins (in order to verify they're instances of the
Hook class) this code is now running before the options from the config
file have been set. Instead, lazily read the options the first time the
key property is accessed, which should be just after the place where
they were previously being constructed.