Use winston logging framework to structure logs
Using winston means we can keep track of which logs are coming from
which hooks/tasks. We also use it to label output coming from spawned
processes and calls to dependencies.

A logger property has been added to the Hook and Task classes so that
the logger can be used from any of their methods (with the name of the
hook/task automatically stored as metadata.) If you need to log in
functions called to by the objects you will need to pass their logger as
a parameter: just using console.log will mean we lose the context. The
no-console lint has been set to 'error', preventing accidental uses of
console.log being committed – we might be want to loosen this to its
default 'warn' value after we've settled into winston.

We now also assign a log level to each logging message (e.g., 'error',
'info', 'verbose'.) All logging below 'info' will be filtered out by
default which should make tasks less noisy whilst still keeping the logs
in place for when we need to debug the code (the ability to dynamically
change the minimum level will be added in a future commit.)