Fix the hookConsole helper function
Thanks to a copy and paste bug where the stderr stream was being hooked
twice and the stdout stream wasn't being hooked at all (but the hooked
stderr streams were being redirected to it) a serious bug was obscured –
the hooked call to winston would just log to the stdout stream and hit
the hook again, causing an infinite loop that would never print.

All we want to do is catch any calls to process.stdout/err, format them
with winston, then pass them on to the original stdout/err streams. This
can get a bit tricky as the winston logger doesn't provide a simple
method to get the message it would normally pass to its transports (its
format property doesn't include the metadata associated with the logger,
and you can't extract that metadata from child loggers as its only
stored in a lambda function.) Instead, we have to create a simple custom
transport that will get the logged message and call a hook defined at
instantiation that will write to the original stdout/err stream.