Only lint staged files in pre-commit hook
Should significantly speed up the time it takes to run the hook, which
would previously often run for upwards of 10 seconds every time a commit
was made. We don't need to run a whole lint each time; during a commit
we only care about verifying the staged files. Other files may have
issues but that's okay as we aren't committing them right now.