chore: release 2.0 version for all packages
(Creating an empty commit will create a release for all packages in the
monorepo.)

Now that we're using release-please to manage our releases we can start
making releases for packages independent of each other. This will mean
we'll start reasoning about and respecting semver conventions. To
demarcate the change in versioning semantics, we're creating a new major
release of all the packages currently in the project. In the future you
should not expect that package versions will be aligned, not even by
their major version. They will still use dependency ranges to indicate
what other packages in the monorepo they are compatible with, however.

release-as: 2.0.0