build: handle premajor semantic versioning properly
These settings mean that breaking changes for premajor versions are
treated as minor bumps, and new features are treating as patch bumps.
This is more in line with how semantic versioning is typically
treated (think how the caret range in npm behaves for 0.x versions.) In
order to release a 1.0 you should include a `release-as: 1.0.0` footer
in the commit message.