build: add a lockfile to the workspace
Now that we aren't using athloi anymore it will be okay to include a
lockfile. Seeing as this is an npm workspace, all the packages in the
project share the same lockfile at the workspace root (along with
sharing a `node_modules` directory)

This should help improve installation performance and make sure
behaviour is consistent between installations.