fix(heroku): scale dynos before promoting them to production
When the staging app is promoted to production, the pipeline will enter
a 'preboot phase' where both the old and new production apps are run in
parallel and, once it is ready, requests will start being routed to the
new app. This is to minimise downtime.

During the preboot phase, however, the formation cannot be changed,
i.e., the app cannot be scaled to a different number of dynos. This
process can take a few minutes so we can't just leave Tool Kit running
waiting for it to make any scaling changes. Instead, let's scale the app
before we promote the new version to production. This will mean that the
old app might be scaled as well but I can't think of any reason this
would be a concern.