feat(mocha): allow plugin to pick up .mocharc files
We want to delegate as much configuration to tools themselves rather
than exposing options via Tool Kit. Unfortunately, the mocha plugin was
invoking mocha with a low-level API and skipping the code which reads
configuration files, meaning that users could not set options like they
can with plugins such as Jest and Webpack.

To solve this let's just invoke the mocha CLI directly rather than via a
lower-level API so that it calls all the setup code it would for a
normal invocation, including looking for configuration files. The user
can also set the path for the config file explicitly in the Tool Kit
options.

The files that mocha reads can now be set in its own configuration file
so we don't really need to have them be set by Tool Kit. The option to
do so might be removed in a future major release of the plugin.