test(cli): remove config snapshot test
We are now also storing the version of the types plugin that Hooks and
Tasks are associated with, meaning that the snapshot won't match after
every update. Of course we could normalise the version field in the
classes the same as we do with file paths, but there's already enough
hacky code trying to hold the snapshot up without any obvious benefits
to keeping the test running. We always update the snapshot whenever we
make any changes to the config that might affect it, and we aren't
making any guarantees to consumers that the structure will be stable.
So let's scrap the snapshot.