feat(pa11y): allow Pa11y to use config defined in .pa11yci
Instead of passing static options to Pa11y, calling the pa11y-ci binary
instead reads config defined in a .pa11yci[.js[on]] file. Seeing as this
can be a .js file, the config can be dynamically generated, making the
tooling much more flexible. This was how n-gage interfaced with Pa11y.

BREAKING CHANGE: this changes the pa11y config structure which users
will have to update to