fix(n-test): set Heroku app name for staging apps too
We want to run smoke tests for both review and staging Heroku apps, so
check for written state for both in the n-test plugin. The state isn't
shared between CI workflows so it should be safe to assume that any
workflow that is staging will not have any review metadata stored.