fix(pa11y)!: remove Pa11y task from staging test hook
The Pa11y task does not currently support reading state from the staging
metadata. Rather than add support I noticed that n-gage never ran Pa11y
when staging for a deployment historically so let's mirror that
behaviour here. If we find that some users consider it useful to be able
to run Pa11y during staging we can add support for reading the staging
metadata, though we still need not assign it to the hook by default,
instead leaving that up to them configuring in their own .toolkitrc.yml.