docs(create): fix core/create changelog for dependency bumps
There seems to be a glitch with release-please where version updates for
components which only consist of dependency bumps are appended to the
top of the CHANGELOG.md file out of order and without a version header.
Let's manually fix this and take the opportunity to add back some commit
messages that were missed by release-please when merging an old branch.

feat(core): gather errors when loading config and throw all together
refactor(core): rename Config to RawConfig