fix(heroku): properly test if app has been deployed previously
We need different deployment behaviour for Heroku apps depending on
whether they've been deployed before or not. Previously, we'd just check
for whether an app existed in the pipeline, but this isn't a useful
heuristic as apps are apparently created before a build is run. So let's
instead call the Heroku API and see what the 'slug size' of an app is:
this value appears to be null until the app has been build and promoted
to production.

This does mean that we're making an extra API call before we deploy an
app, but this should hopefully result in more reliable deployments.