fix(upload-assets-to-s3): strip whole base directory from upload key
Tool Kit has a plugin to upload assets within a folder to S3. You can
set a directory option to specify where in your project the public
assets are stored. The plugin is meant to strip out the directory part
of the path from the destination key so that you aren’t forced to have a
public/assets/blah/ folder structure in S3 that mirrors your project’s
directory structure. However, previously, only the first directory path
is stripped from the key, so files would still have a path like
assets/blah/foo.txt.

I think this is safe to release as a patch instead of a breaking change
because the previous behaviour was surprising and I don't think anyone
was relying on it.