fix(cli): avoid infinite loop when custom plugin doesn't have config
We've found that it's quite easy to forget to create a .toolkitrc.yml
config file for your custom plugin (even though it's mentioned in the
documentation). Omitting the config file will mean that cosmiconfig will
try and traverse parent directories until it finds a valid config. This
will likely be the config at the app's root instead, which will
(at least transitively) include the current plugin in its `plugin`
property, causing an infinite loop of config resolution.

Let's try and catch when we've resolved the config to the same root
version and print a warning to remind people this can have unexpected
results. This probably won't catch loops when one plugin installs
another within the `node_modules` directory, but at that point it's
likely going to be a plugin we own and have audited so should have a
config file.