feat(cli): allow state to be shared between install hooks
Previously, every install method in a hook implementation would have to
run independently of every other hook, even if other hooks would be
writing to the same file. This was very complicated to manage and
grossly inefficient, as, for instance, every CircleCI hook had to read
and write to the same .circleci/config.yml file, one after the other.
Some hooks also added dependencies on other hooks: if one CircleCI hook
adds a filter to run on git tags, then the other hooks needed to follow
suit so that they weren't ignored for a tag release. This meant a
hardcoded postInstall function was required in the CLI to handle some
post-processing depending on which hooks ran.

Now, state is able to be shared between hooks. A state object is passed
between hooks who have the same installGroup property, e.g., all
CircleCI hooks have the 'circleci' installGroup. If state was set
between hooks, then an additional commitInstall method is called after
all the hooks in the group have run which will actually commit the
generated state, e.g., write a .circleci/config.yml file based on the
state.

I've tried to make these changes as backwards compatible as possible to
reduce developer friction, and plugins using the old style of hooks
should be compatible with these changes. However, updating to the new
style of hooks will be a breaking change for a plugin.