feat(package-json-hook)!: share state for package.json hook installers
This is a breaking change because these new install hooks rely on the
new commitInstall method to actually write to your package.json file,
and will conflict with the older implementation which rewrote the
manifest file for every hook. If some of your hooks use the old version
of this plugin, and others use the new version, then any hooks using the
old version will be lost, as commitInstall is always called last and
ignores any hooks not present in its state object.