refactor(circleci): simplify code and replace js-yaml with yaml package
Move some duplicated logic into a shared function, eliminate a
duplicated class field, and use the yaml package. The latter is to keep
consistency with the rest of Tool Kit which also uses the yaml package,
and also allows us to drop down into lower level YAML manipulation if
necessary.