feat(circleci)!: share state between circleci install hooks
This is a breaking change because these new install hooks rely on the
new commitInstall method to actually write to you .circleci/config.yml
file, and will conflict with the older implementation which rewrote the
config file for every hook. If some of your hooks use the old version of
this plugin, and others use the new version, then any hooks using the
old version will be lost, as commitInstall is always called last and
ignores any hooks not present in its state object.