fix(create): handle npm installation failing
The create package has some special handling for when the error being
thrown is relating to Tool Kit conflicts. To check this we need to pull
in a package defining the types from the local project rather than from
a local Tool Kit installation (or a different version from the
registry), as otherwise some of the instanceof checks won't work. So we
import our error package before printing any errors.

However, sometimes that package won't be available as something failed
before we could even finish an npm instllation. In those cases it's safe
to assume that the issue isn't related to tool kit conflicts as we
haven't even installed tool kit yet, so if we can't import the errors
package let's assert this isn't related to conflicts. This allows us to
print the actual error, rather than obscuring it by an error about the
error package being missing.