fix(create): clear cache from correct cosmiconfig object
The migration script will try and install your config and check for
hook conflicts. If there are any conflicts it will allow you to specify
the order of the ambiguous tasks, then try and run the installation
again. This second installation has been failing consistently as it
doesn't seem to be finding the newly configured task order in the
config. This seemed strange as we were already explicitly clearing the
cache of the cosmiconfig object we use to load the config files.

It turns out the cache wasn't actually being cleared as expected because
we've been using the cosmiconfig object installed into the target
project's node_modules rather than the one from wherever npx chooses to
install the script since fcd2a7501c9ee2e670f489cd2a1d0364418260d5. Let's
instead get the cosmiconfig object via importCwd() so that we can
control the config cache properly.