fix(create): be explicit if second installation attempt fails
The migration script checks for hook conflicts when it intalls Tool Kit,
asks how you want to resolve them if there are any, then runs the
installation again. In theory, this second installation should never
result in conflicts, as we've made sure to handle all the ambiguous
tasks raised by the first implentation explicitly. However, naturally,
if there is a bug in this script then it is still posible for this
assumption to fail to be upheld.

There's some special logic in the script's error handling to check is
the error is about hook conflicts and not log the installation task as
having failed if that's the error that was thrown. This makes it look
like the code executed successfully if the installation fails on the
second attempt, which will be confusing for an end user as it will look
like they are done with the migration (when in fact they haven't been
able to set up their config options yet) unless they are checking the
exit status of the script carefully. Instead, let's add a switch to the
logging function to amke the logging to print out the error and mark the
execution as failed if we're still getting hook conflicts on the second
attempt.