fix(circleci-heroku): run provision step even for nori/renovate builds
I think this was always an oversight rather than something we can only
do now that we've made other improvements to the code. The job now
requires the 'waiting-for-approval' job to have been run first, though
this only runs for nori/renovate branches. If that job is filtered out
then this job will instead simply run after 'tool-kit/setup' is
completed. n-gage would previously create a different workflow for
nori/renonate, with the 'waiting-for-approval' job required before
anything else was run. Tool Kit's approach is much simpler and reduces
duplication.