feat(circleci): add new error wording to circleci plugin
to provide direction on generating automated config

currently, if a user is missing specific Tool Kit jobs in their config, a tailored
error is thrown with the hooks they need to install. However, if a user is missing
those hooks because their config is legacy config that hasn't yet been moved to
Tool Kit (e.g. because they are migrating to Tool Kit for the first time), it is
better to throw an error that explains how to auto generate Tool Kit config in the
first place