feat(create): add migration path for scheduled workflows
CircleCI has deprecated scheduled workflows and have replaced them with
scheduled pipelines, and Tool Kit consequently only supports the latter.
We found that many teams were missing this change during migration and
as a result no longer have their nightly workflow running every night.
To avoid this in the future let's highlight that the migration is
necessary in the migration script, as well as offer to set up the new
workflow programmatically via the CircleCI API.