build: bump release-please to v15
We are jumping from v13 to v15. v14 was breaking because it dropped
support for Node 12, which we don't use, and v15, which just released
yesterday, and doesn't seem to make any breaking changes that would
affect us.

We previously had an issue where newer versions of release-please would
be incorrectly trying to create releases for packages that didnt; need
updating. I tested to see whether this applied to v15 too and couldn't
reproduce the issue, though we currently already have 29 updates due.