fix(upload-assets-to-s3): only upload files
the plugin works by looping over the files passed to it and uploading each to s3.
When the files to upload includes a folder (with files inside)
the directory itself is passed to s3 for upload.
s3 throws an error because it can only read files and not directories.
This change updates the glob pattern of files to upload to not include directories.