chore: bump release-please bootstrap SHA
This is the commit ID that is the furthest back release-please will
search for releases. Normally, you can just set it and forget it, as
subsequent releases will automatically use the previous release as the
baseline rather than the bootstrap commit, but this bootstrap commit is
still being used when we create a new plugin, for instance, that doesn't
have a previous release.

This was causing trouble as when we switched to release-please we forced
every component to v2.0.0 to delineate them from the old versioning
strategy. That commit was made since the original bootstrap commit and
was being picked up by release-please when trying to work out the
version for a new component. This meant that each new plugin was
erroneously being versioned as v2.0.0 rather than an initial v1.0.0
release.

The new bootstrap commit is one of the parent commits of the merge
commit for the last release-please PR. This means that the next release
will consider all the commits up to the last release – which is how the
typical behaviour anyway – but new components won't pick up the old
v2.0.0 commit.