feat(types): use Zod for option schemas
Previously, we'd rolled our own method of specifying option schemas
using strings and complicated TypeScript magic to infer equivalent
parsed types, but the code was quite hard to maintain and not
particularly extensible. Instead let's use the TS parsing and validation
library, Zod, to do the job for us.

The main advantage we got from just using string literals was that
reflection of what the types were in JS code was pretty simple. In
theory, you could use instanceof to inspect what types each class in a
parsing library were associated with, but this would be quite fragile as
instanceof can't handle classes from different versions of a dependency
so there would need to be some version pinning between packages in the
monorepo. Fortunately, Zod uses an enum in its parser definitions that
we can inspect and that will be compatible between versions. It doesn't
seem to be undocumented and may not be stable but its better than the
comparison breaking after each patch update.