fix(upload-assets-to-s3): better error handling on looped promises
Fixes the scenario where the task does not return an error
even when it has not executed successfully

The failure scenario is:

Two buckets are specified for upload (e.g. ['ft-next-service-registry-prod',
'ft-next-service-registry-prod-us'])
AND
Uploads are successful to the first bucket
AND
There is an error in uploads to the second bucket

this issue is caused due to with looping with promises - a race condition where
resolved and rejected promises are returning unpredictably and interfering with each
other.

The solution is to wrap all the promises in a Promise.all and then handle the
successful responses once they've all resolved.