build: avoid importing whole DOM TS library
aws-sdk v3 is designed to be an isomorphic library, supporting both Node
and Web APIs. This means some of its type definitions use DOM types that
we don't use, causing type errors. But we would like to avoid pulling in
the whole global DOM type namespace to fix this, so instead just declare
a skeleton type for the problematic Blob type.