test(circleci): remove redundant test
All this test did was check that the circleci module was exporting its
hooks. It's trivial to see whether the `hooks` export has been changed
or removed and this test mostly just creates noise if we do decide to
change the exported hooks (and doesn't account for any additional hooks
we might add later).