feat(circleci)!: rewrite CircleCI plugin to allow more flexible hooks
When first writing this plugin's .install() behaviour, I found that
defining the shape of a CircleCI config file could be quite complicated,
but that there were some general patterns which we could abstract into
something simpler. We initially only allowed hooks implementing the
CircleCI installation code to specify a single job, but soon found this
to be too limiting. Some hooks require additional options on other jobs,
or need to specify options outside of a workflow, or need to declare a
new job etc. We kept piling on more abstractions and increasing the code
complexity to the point where it was fairly unmaintainable. Now, we need
to support a hook that can install multiple jobs so that OIDC
authentication can work for the Serverless plugin, and rather that
adding another pile of straw, I've resolved to rewrite the architecture
instead.

Now, each hook specifies a partial view of the CircleCI configuration,
declaring the parts the hook is concerned with, e.g., a new job in the
tool-kit and nightly workflows. These partial views are then merged
together on top of a default configuration. However, each partial view
needs to define the whole nested configuration structure preceding the
changes it's concerned with (the path to a new job is
`workflows["tool-kit"].jobs[].[jobName]` for instance), and there is a
lot of repetitive boilerplate each job needs but hooks needn't be
concerned with. Therefore, a generator function is available that allows
you to specify a job's parameters and get the partial view in return. It
is then relatively easy to make further arbitrary modifications to this
partial view via object spreads if required.

This new approach allows for any hook to make arbitrary changes to the
CircleCI config that don't require any new code support in the circleci
plugin. I've also taken the chance to remove the 'postinstall' logic
which would previously make further modifications to the config object
after all hooks had finished specifying their jobs etc. This would,
before these changes, set up tag filters if tagged releases were used,
and use a custom executor if we wanted to override the docker image or
tag. I've found that both these tasks can just be done unconditionally
regardless of your Tool Kit options and have consequently removed the
'postinstall' logic that checks if certain hooks have been used. This
gives a couple of benefits: 1) it makes the code much simpler, and 2) it
reduces the number of assumptions we have to make about the structure
of the computed config when running .commitInstall(). It does result in
configuration files that are more verbose and contain some settings that
may be redundant but I think the benefits outweigh the costs in this
case.