refactor(circleci): use static field for job name
Now that the abstract CircleCiConfigHook class no longer references a
`job` field we can just define that field statically for each subclass
so that they can reference each other's job names when specifying
requirements without having to instantiate the required class. This is
still not possible for subclasses of CypressCiHook, as that abstract
class still uses the `job` field, and abstract static fields are not
currently possible in TypeScript.