build: force the same version of type-fest across packages
type-fest was originally added to the circleci package to make use of
its PartialDeep type for the abstract config field in
CircleCiConfigHook, which was then implicitly reused by other packages
that overrode the property. TypeScript threw errors about this package
being used implicitly, which I haven't seen before, and is surely
something that is relatively common in the monorepo? Well, it seems that
pinning the package version across the monorepo seems to fix the issue
though – the version that was hoisted to root wasn't
@dotcom-tool-kit/circleci's version so previously other packages were
using the types from a different major version.