feat(orb): add support for assuming AWS role in provisioning job
I was initially reticent to put this serverless logic in the
heroku-provision job (whose name is still needing an update) as it kind
of breaks the generic design of the Tool Kit hooks. However, after
experimenting with trying to put the AWS assume role logic in a
serverless Tool Kit hook, I found that the AWS logic is assumed to all
run in a single job, as the AWS credentials are stored in a file that
is not manually persisted through workspaces between jobs by default.
Workarounds are possible by writing environment variables to override
the credentials file's location to $BASH_ENV, but this approach creates
verbose and confusing CircleCI configuration files.

Instead, the pragmatic solution seems to be to put the logic in the
Tool Kit orb's job, but only execute the AWS command if you pass the
parameters it requires to the job, thereby allowing Heroku apps to use
the job with no change in logic. This was inspired by CircleCI's aws-cli
orb which does something similar when decided whether to assume a role
or use credentials passed to it directly.