refactor: decouple the Options type from the options library
The options library exports a function to read and write to the global
options object build from the options in Tool Kit configurations. This
object doesn't have to contain options for every plugin we define, i.e.,
it's a partial object. It was pulling in the type from the types
library, but this type was written to be partial itself solely to suit
the options library. There's no need for these to be coupled in reality,
and it's trivial to make the object partial in the options library by
using the Partial type.