chore: require latest cli version for all plugins
The latest version of @dotcom-tool-kit/cli adds support for zod schemas.
In doing so, we moved the handling of default values for options from
a line in the Task constructor to using zod's built-in support for
default values during parsing. This means that if we do not use zod's
options parsing then we are no longer setting default values for
options. To avoid this we need to bump the range that plugins allow for
the dotcom-tool-kit package peer dependency to make sure we're
definitely installing a version with support for zod. I wouldn't
consider this a breaking change any more than bumping any Tool Kit
dependency is.