fix(husky-npm): add husky as a dependency
When a repo installs to @dotcom-tool-kit/husky-npm we expect husky to be
bundled so that our pre-commit hook can execute. However it turns out
that when husky is defined as a peer dependency, although it does get
installed, its executable is not defined in the root node_modules/.bin
folder. This means that the `husky-run` executable cannot be found by
the git pre-commit hook.

The plugin should depend on husky. This matches the way it worked for
n-gage.