fix(create): set options before installing to avoid validation errors
Previously, we would try installing your Tool Kit config (to test for
plugin conflicts) before prompting for option values. This was so that
we could get the loaded config that the installation process returns,
which we could use to get a list of all the plugins used in your app
so we only need to prompt for options for those specific plugins.
However, this approach started causing issues when we add validation for
options via zod. As some plugins must have their options defined before
they are considered valid, the installation validation would fail
because we hadn't set these options yet.

Now, we prompt the user for options before the installation (and
conflicts check) is run. We can do this by simply loading the config
with validation disabled (an option we already supported) instead of
getting the config from a full install. We can then run the install
after the options have been filled out. This has the extra benefit that
some installation logic is actually dependent on options you specify,
so sometimes users would have had to manually rerun the installation
previously.

This subsequent reordering of the logic also meant that some of the
extended komatsu logger functionality we were using is no longer
necessary (we were showing we were waiting for the installation to begin
whilst some prerequisite tasks were run, but now the installation is run
at a different time with only a single trivial file write task happening
first). This meant I could strip out a lot of the custom logger logic,
including no longer subclassing the komatsu logger class, so that the
logger module is now a lot simpler.