feat(create): syntax highlight the AWS OIDC YAML config
This makes it much easier for a user to parse what is going to be
included in a PR to an aws-composer repository. I don't love the
selection of syntax highlighters that are available: all of them use the
highlight.js library which is massive, and I'd rather a library that
used the much smaller prismjs existed. A more modern library, emphasize,
uses ESM modules so we can't use it here :( The default theme for
cli-highlight isn't super pretty but maybe I'm just used to my nice
neovim and bat themes. Maybe we can revisit this later if we find a more
lightweight way of highlighting some YAML source.