fix(create): load config via user's local dotcom-tool-kit package
This was causing a crash as dotcom-tool-kit is only a dev dependency of
the create package. Rather than promoting it to a full dependency to use
loadConfig, I remembered that in other places in the package we use the
version of dotcom-tool-kit that we install into the user's own project
as part of the migration to avoid instanceof incompatibilities. It makes
sense to keep doing the same thing here to avoid accidentally bumping
into further incompatibilities in future code changes.